# Conflicts:
#	src/SixLabors.Fonts/Tables/Cff/FDRangeProvider.cs
#	src/SixLabors.Fonts/Tables/TableLoader.cs

+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/NotImplementedSubTable.cs
@@ -15,7 +15,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GSub
             GSubTable table,
             GlyphSubstitutionCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count)
             => false;
     }
